pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code-review-env"
version = "0.1.0"
description = "OpenEnv RL environment for AI-powered code review"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
inference = [
    "openai>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["server*", "models*"]
